Remove old Mach-O parser, libhelper-macho-ll now becomes libhelper-macho
