+++ include/libhelper-macho/macho-header-const.h
@@ -1,21 +1,27 @@
-/**
- *     libhelper
- *     Copyright (C) 2019, @h3adsh0tzz
- *
- *     This program is free software: you can redistribute it and/or modify
- *     it under the terms of the GNU General Public License as published by
- *     the Free Software Foundation, either version 3 of the License, or
- *     (at your option) any later version.
- *
- *     This program is distributed in the hope that it will be useful,
- *     but WITHOUT ANY WARRANTY; without even the implied warranty of
- *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *     GNU General Public License for more details.
- *
- *     You should have received a copy of the GNU General Public License
- *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
-*/
+//===-----------------------  macho-header-const ----------------------===//
+//
+//                          Libhelper Mach-O Parser
+//
+//  This program is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  This program is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU General Public License for more details.
+//
+//  You should have received a copy of the GNU General Public License
+//  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+//
+//
+//  Copyright (C) 2019, Is This On?, @h3adsh0tzz
+//  me@h3adsh0tzz.com.
+//
+//
+//===------------------------------------------------------------------===//
+
 /*
  * Copyright (c) 1999-2010 Apple Inc.  All Rights Reserved.
  *
@@ -39,8 +45,8 @@
  * @APPLE_LICENSE_HEADER_END@
  */
 
-#ifndef LIBHELPER_MACHO_HEADER_CONST_H
-#define LIBHELPER_MACHO_HEADER_CONST_H
+#ifndef LIBHELPER_MACHO_HEADER_CONST_LL_H
+#define LIBHELPER_MACHO_HEADER_CONST_LL_H
 
 /**
  *                  === The Libhelper Project ===
@@ -201,4 +207,4 @@
 #define     MH_DYLIB_IN_CACHE           0x80000000
 
 
-#endif /* libhelper_macho_header_const_h */
+#endif /* libhelper_macho_header_const_ll_h */
