+++ src/macho/macho-segment.c
@@ -1,27 +1,34 @@
-/**
- *     libhelper
- *     Copyright (C) 2019, @h3adsh0tzz
- *
- *     This program is free software: you can redistribute it and/or modify
- *     it under the terms of the GNU General Public License as published by
- *     the Free Software Foundation, either version 3 of the License, or
- *     (at your option) any later version.
- *
- *     This program is distributed in the hope that it will be useful,
- *     but WITHOUT ANY WARRANTY; without even the implied warranty of
- *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *     GNU General Public License for more details.
- *
- *     You should have received a copy of the GNU General Public License
- *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
-*/
+//===--------------------------- macho_segment ------------------------===//
+//
+//                          Libhelper Mach-O Parser
+//
+//  This program is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  This program is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU General Public License for more details.
+//
+//  You should have received a copy of the GNU General Public License
+//  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+//
+//
+//  Copyright (C) 2019, Is This On?, @h3adsh0tzz
+//  me@h3adsh0tzz.com.
+//
+//
+//===------------------------------------------------------------------===//
 
 #include "libhelper-macho/macho-segment.h"
+#include "libhelper-macho/macho.h"
 
-//////////////////////////////////////////////////////////////////////////
-//               Base Mach-O Segment commands                           //
-//////////////////////////////////////////////////////////////////////////
+
+//===-----------------------------------------------------------------------===//
+/*-- Mach-O Segments                     									 --*/
+//===-----------------------------------------------------------------------===//
 
 
 /**
@@ -38,15 +45,16 @@ mach_segment_command_64_t *mach_segment_command_create ()
 /**
  * 
  */
-mach_segment_command_64_t *mach_segment_command_load (file_t *file, off_t offset)
+mach_segment_command_64_t *mach_segment_command_load (unsigned char *data, uint32_t offset)
 {
     mach_segment_command_64_t *sc = mach_segment_command_create ();
-    sc = (mach_segment_command_64_t *) file_load_bytes (file, sizeof(mach_segment_command_64_t), offset);
+    memset (sc, '\0', sizeof (mach_segment_command_64_t));
+    memcpy (sc, data + offset, sizeof (mach_segment_command_64_t));
 
     if (!sc) {
         debugf ("[*] Error: Problem loading Mach Segment Command at offset 0x%llx\n", offset);
         exit (0);
-    } 
+    }
 
     return sc;
 }
@@ -57,96 +65,45 @@ mach_segment_command_64_t *mach_segment_command_load (file_t *file, off_t offset
  */
 mach_segment_info_t *mach_segment_info_create ()
 {
-    mach_segment_info_t *si = malloc (sizeof (mach_segment_info_t));
-    memset (si, '\0', sizeof(mach_segment_info_t));
-    return si;
+    mach_segment_info_t *ret = malloc (sizeof (mach_segment_info_t));
+    memset (ret, '\0', sizeof (mach_segment_info_t));
+    return ret;
 }
 
 
 /**
  * 
  */
-mach_segment_info_t *mach_segment_info_load (file_t *file, off_t offset)
+mach_segment_info_t *mach_segment_info_load (unsigned char *data, uint32_t offset)
 {
-    // Create a new segment info struct and load the segment command from file
-    mach_segment_info_t *si = mach_segment_info_create ();
-    mach_segment_command_64_t *segment = (mach_segment_command_64_t *) file_load_bytes (file, sizeof(mach_segment_command_64_t), offset);
+    // Create a new segment info struct and load the segment command
+    mach_segment_info_t *seg_inf = mach_segment_info_create ();
+    mach_segment_command_64_t *segment = mach_segment_command_load (data, offset);
 
-    // Check that a segment was correctly loading from the offset at the file
+    // Check that the segment cmmmand is valid
     if (!segment) {
-        debugf ("[*] Error: Could not load Segment (64) at offset 0x%llx\n", offset);
-        exit (0);
+        errorf ("Could not load Segment Command\n");
+        return NULL;
     }
 
-    // Fiddle with the vmaddr to ensure non-kernel images are loading correctly
-    if (segment->vmaddr < 0xffffff0000000000) {
-        uint64_t vmaddr = 0xffffff0000000010 + segment->vmaddr;
-        segment->vmaddr = vmaddr;
-        si->padding = 0xffffff0000000000;
-    } else {
-        si->padding = 0x0;
-    }
-
-    // Calculate the offset and start loading section commands
-    offset = segment->vmaddr;
+    //  the section commands are placed directly after the segment command.
+    uint32_t sectoff = offset + sizeof (mach_segment_command_64_t);
     for (int i = 0; i < (int) segment->nsects; i++) {
 
-        // Load a new section 64.
-        mach_section_64_t *sect = mach_section_load (file, offset);
-
-        // Append the section to the sections list of the segment info struct
-        si->sections = h_slist_append (si->sections, sect);
-
-        // Increment the offset
-        offset += sizeof (mach_section_64_t);
+        // Load a section 64
+        mach_section_64_t *sect = mach_section_load (data, sectoff);
+    
+        seg_inf->sections = h_slist_append (seg_inf->sections, sect);
+        sectoff += sizeof (mach_section_64_t);
     }
 
-    // Set the segment command for the info struct
-    si->segcmd = segment;
-
-    // Return that segment info
-    return si;
+    seg_inf->segcmd = segment;
+    return seg_inf;
 }
 
 
 /**
  * 
- */
-mach_segment_info_t *mach_segment_command_search (macho_t *mach, char *segname)
-{
-    // Check the segname given is valid
-    if (!segname) {
-        debugf ("[*] Segment name not valid\n");
-        exit (0);
-    }
-
-    // Get the amount of segment commands and check its more than 0
-    int c = h_slist_length (mach->scmds);
-    if (!c) {
-        debugf ("[*] Error: No Segment Commands\n");
-        exit (0);
-    }
-
-    // Now go through each of them
-    for (int i = 0; i < c; i++) {
-        
-        // Grab the segment info
-        mach_segment_info_t *si = (mach_segment_info_t *) h_slist_nth_data (mach->scmds, i);
-        mach_segment_command_64_t *s = si->segcmd;
-
-        // Check if they match
-        if (!strcmp(s->segname, segname)) {
-            return si;
-        }
-    }
-
-    // Output an error
-    debugf ("[*] Could not find Segment %s\n", segname);
-    return NULL;
-}
-
-
-/**
  * 
  */
 HSList *mach_segment_get_list (macho_t *mach)
@@ -173,48 +130,43 @@ HSList *mach_segment_get_list (macho_t *mach)
 /**
  * 
  */
-mach_segment_info_t *mach_segment_find_with_name (macho_t *macho, char *segname)
+mach_segment_info_t *mach_segment_command_search (HSList *segments, char *segname)
 {
-    for (int i = 0; i < h_slist_length (macho->scmds); i++) {
-        mach_segment_info_t *inf = (mach_segment_info_t *) h_slist_nth_data (macho->scmds, i);
-        mach_segment_command_64_t *segment = inf->segcmd;
+    // Check the segname given is valid
+    if (!segname) {
+        debugf ("[*] Segment name not valid\n");
+        exit (0);
+    }
 
-        if (!strcmp(segment->segname, segname)) {
-            return inf;
+    // Get the amount of segment commands and check its more than 0
+    int c = h_slist_length (segments);
+    if (!c) {
+        debugf ("[*] Error: No Segment Commands\n");
+        exit (0);
+    }
+
+    // Now go through each of them
+    for (int i = 0; i < c; i++) {
+        
+        // Grab the segment info
+        mach_segment_info_t *si = (mach_segment_info_t *) h_slist_nth_data (segments, i);
+        mach_segment_command_64_t *s = si->segcmd;
+
+        // Check if they match
+        if (!strcmp(s->segname, segname)) {
+            return si;
         }
     }
+
+    // Output an error
+    debugf ("[*] Could not find Segment %s\n", segname);
     return NULL;
 }
 
 
-/**
- * 
- */
-void mach_segment_command_dump (mach_segment_info_t *si)
-{
-    mach_segment_command_64_t *sc = si->segcmd;
-
-    if (!sc) {
-        debugf ("[*] Error: Segment command not loaded properly!\n");
-        exit (0);
-    }
-
-    debugf ("Command:\t\t%s\n", mach_load_command_get_string ((mach_load_command_t *) sc));
-    debugf ("Segment Name:\t\t%s\n", sc->segname);
-    debugf ("VM Address:\t\t0x%llx\n", sc->vmaddr);
-    debugf ("VM Size:\t\t%llu\n", sc->vmsize);
-    debugf ("File Offset:\t\t0x%llx\n", sc->fileoff);
-    debugf ("File Size:\t\t%llu\n", sc->filesize);
-    debugf ("Max VM Protection:\t%d\n", sc->maxprot);
-    debugf ("Initial VM Protection:\t%d\n", sc->initprot);
-    debugf ("Number of Sections:\t%d\n", sc->nsects);
-    debugf ("Flags:\t\t\t%d\n\n", sc->flags);
-}
-
-
-//////////////////////////////////////////////////////////////////////////
-//               Base Mach-O Section commands                           //
-//////////////////////////////////////////////////////////////////////////
+//===-----------------------------------------------------------------------===//
+/*-- Mach-O Sections                     									 --*/
+//===-----------------------------------------------------------------------===//
 
 
 /**
@@ -222,46 +174,24 @@ void mach_segment_command_dump (mach_segment_info_t *si)
  */
 mach_section_64_t *mach_section_create ()
 {
-    mach_section_64_t *s = malloc(sizeof(mach_section_64_t));
-    memset (s, '\0', sizeof(mach_section_64_t));
-    return s;
-}
-
-
-/**
- * 
- */
-mach_section_64_t *mach_section_load (file_t *file, off_t offset)
-{
-    mach_section_64_t *s = NULL; //mach_section_create ();
-    s = (mach_section_64_t *) file_load_bytes (file, sizeof(mach_section_64_t), offset);
-
-    if (!s) {
-        debugf ("[*] Error: Problem loading section at offset 0x%llx\n", offset);
-        exit (0);
-    }
-
-    return s;
-}
-
-
-/**
- * 
- */
-HSList *mach_sections_load_from_segment (macho_t *macho, mach_segment_command_64_t *seg)
-{
-    HSList *ret = NULL;
-    uint64_t offset = seg->vmaddr;
-    
-    for (int i = 0; i < (int) seg->nsects; i++) {
-        debugf ("Loading %lu bytes from 0x%llx\n", sizeof(mach_section_64_t), offset);
-        mach_section_64_t *sect = (mach_section_64_t *) file_load_bytes (macho->file, sizeof(mach_section_64_t), offset);
-        ret = h_slist_append (ret, sect);
-
-        offset += sizeof(mach_section_64_t);
-    }
-
+    mach_section_64_t *ret = malloc (sizeof(mach_section_64_t));
+    memset (ret, '\0', sizeof(mach_section_64_t));
     return ret;
+}
+
+
+/**
+ * 
+ */
+mach_section_64_t *mach_section_load (unsigned char *data, uint32_t offset)
+{
+    mach_section_64_t *sect = mach_section_create ();
+    memcpy (sect, data + offset, sizeof (mach_section_64_t));
+
+    if (sect == NULL) {
+        errorf ("There was a problme loading the section at offset 0x%x\n");
+    }
+    return sect;
 }
 
 
@@ -296,9 +226,8 @@ mach_section_64_t *mach_search_section (mach_segment_info_t *info, char *sectnam
 /**
  * 
  */
-mach_section_64_t *mach_find_section (macho_t *macho, int sect)
+mach_section_64_t *mach_find_section (HSList *segments, int sect)
 {
-    HSList *segments = macho->scmds;
     int count = 0;
     for (int i = 0; i < h_slist_length (segments); i++) {
         mach_segment_info_t *seg = (mach_segment_info_t *) h_slist_nth_data (segments, i);
@@ -320,33 +249,16 @@ mach_section_info_t *mach_load_section_data (macho_t *macho, char *segment, char
 {
     mach_section_info_t *ret = malloc (sizeof(mach_section_info_t));
 
-    mach_segment_info_t *seginfo = mach_segment_command_search (macho, segment);
+    mach_segment_info_t *seginfo = mach_segment_command_search (macho->scmds, segment);
     mach_section_64_t *__sect = mach_search_section (seginfo, section);
 
     ret->segment = __sect->segname;
     ret->section = __sect->sectname;
-    ret->data = file_load_bytes (macho->file, __sect->size, __sect->offset);
     ret->size = __sect->size;
 
+    ret->data = malloc (__sect->size);
+    memset (ret->data, '\0', __sect->size);
+    memcpy (ret->data, macho->data + __sect->offset, __sect->size);
+
     return ret;
-}
-
-
-/**
- * 
- */
-void mach_section_print (mach_section_64_t *section)
-{
-    debugf ("Section:\t%s\n", section->sectname);
-    debugf ("Segment:\t%s\n", section->segname);
-    debugf ("Address:\t0x%llx\n", section->addr);
-    debugf ("Size:\t\t%llu\n", section->size);
-    debugf ("Offset:\t\t0x%x\n", section->offset);
-    debugf ("Align:\t\t%u\n", section->align);
-    debugf ("Reloff:\t\t0x%x\n", section->reloff);
-    debugf ("Nreloc:\t\t%u\n", section->nreloc);
-    debugf ("Flags:\t\t%u\n", section->flags);
-    debugf ("Indr Sym Index:\t%d\n", section->reserved1);
-    debugf ("Reserved 2:\t%d\n", section->reserved2);
-    debugf ("Reserved 3:\t%d\n\n", section->reserved3);
 }
