+++ src/meson.build
@@ -9,14 +9,9 @@
 #   Sources
 #
 mach_parser_sources =  ['macho/macho.c',
-                        'macho/macho-header.c',
                         'macho/macho-command.c',
                         'macho/macho-segment.c',
                         'macho/macho-symbol.c']
-
-mach_ll_parser_sources =   ['macho-ll/macho.c',
-                            'macho-ll/macho-command.c',
-                            'macho-ll/macho-segment.c']
 
 img4_sources = ['img4/sep.c']
 
@@ -35,8 +30,7 @@ lib_sources = ['libhelper.c',
                 'strutils.c', 
                 'hslist.c', 
                 'hstring.c', 
-#                mach_parser_sources,
-                mach_ll_parser_sources, 
+                mach_parser_sources,
                 img4_sources,
                 lzss_sources,
                 lzfse_sources]
